requirements.txt
Django>=4.2.0
Pillow>=9.0.0
django-crispy-forms>=2.0
crispy-bootstrap5>=0.7
python-decouple>=3.6
openai>=1.0.0
requests>=2.31.0
